LPD-99371 Add an Add Asset button and simplify the assets table
